requirements.txt
python-telegram-bot[job-queue]==21.6
openpyxl==3.1.5
pytz==2024.2
